feature-6/ move user dtos to request and response packages
