+++ src/main/java/jewellery/inventory/controller/UserController.java
@@ -3,8 +3,8 @@ package jewellery.inventory.controller;
 import jakarta.validation.Valid;
 import java.util.List;
 import java.util.UUID;
-import jewellery.inventory.dto.UserRequest;
-import jewellery.inventory.dto.UserResponse;
+import jewellery.inventory.dto.request.UserRequest;
+import jewellery.inventory.dto.response.UserResponse;
 import jewellery.inventory.service.UserService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
